+++ src/schemas/definitions/ArraySchema.ts
@@ -1,13 +1,23 @@
-import type { ParseContext } from '../../types.js';
+import type { ParseContext, Simplify } from '../../types.js';
 import { Schema } from '../Schema.js';
+import { UnionSchema } from './UnionSchema.js';
 
-export class ArraySchema<T> extends Schema<T[]> {
+export type InferArray<T extends [Schema, ...Schema[]]> =
+    Simplify<
+        T extends [infer S]
+            ? S extends Schema<infer U> ? U[] : never
+            : T extends [Schema, ...Schema[]]
+                ? (T[number] extends Schema<infer U> ? U : never)[]
+                : never
+    >;
+
+export class ArraySchema<T extends [Schema, ...Schema[]]> extends Schema<InferArray<T>> {
     constructor(
-        private readonly elementSchema: Schema<T>,
+        private readonly elementSchemas: T,
     ) { super(); }
 
-    protected _normalize(input: unknown, ctx: ParseContext): T[] {
-        let array;
+    protected _normalize(input: unknown, ctx: ParseContext): InferArray<T> {
+        let array: unknown[];
         if (Array.isArray(input)) {
             array = input;
         } else {
@@ -18,16 +28,23 @@ export class ArraySchema<T> extends Schema<T[]> {
                 expected: 'array',
                 received: input,
             });
-
             array = (input === null || input === undefined) ? [] : [input];
+        }
+
+        let normalizer: (item: unknown, ctx: ParseContext) => unknown;
+        if (this.elementSchemas.length === 1) {
+            normalizer = (item, ctx) => this.invokeNormalize(this.elementSchemas[0], item, ctx);
+        } else {
+            const union = new UnionSchema(this.elementSchemas as [Schema, ...Schema[]]);
+            normalizer = (item, ctx) => this.invokeNormalize(union, item, ctx);
         }
 
         return array.map((item, index) => {
             ctx.path.push(index);
-            const result = this.invokeNormalize(this.elementSchema, item, ctx);
+            const result = normalizer(item, ctx);
             ctx.path.pop();
 
             return result;
-        });
+        }) as InferArray<T>;
     }
 }